+++ backend/accounts.go
@@ -50,9 +50,7 @@ func sortAccounts(accounts []*config.Account) {
 			coinpkg.CodeLTC:   2,
 			coinpkg.CodeTLTC:  3,
 			coinpkg.CodeETH:   4,
-			coinpkg.CodeTETH:  5,
-			coinpkg.CodeRETH:  6,
-			coinpkg.CodeGOETH: 7,
+			coinpkg.CodeGOETH: 6,
 		}
 		order1, ok1 := order[coin1]
 		order2, ok2 := order[coin2]
@@ -86,10 +84,12 @@ func (backend *Backend) filterAccounts(accountsConfig *config.AccountsConfig, fi
 	var accounts []*config.Account
 	for idx := range accountsConfig.Accounts {
 		account := &accountsConfig.Accounts[idx]
-		if _, isTestnet := coinpkg.TestnetCoins[account.CoinCode]; isTestnet != backend.Testing() {
-			// Don't load testnet accounts when running normally, nor mainnet accounts when running
-			// in testing mode
-			continue
+		if !backend.arguments.Regtest() {
+			if _, isTestnet := coinpkg.TestnetCoins[account.CoinCode]; isTestnet != backend.Testing() {
+				// Don't load testnet accounts when running normally, nor mainnet accounts when running
+				// in testing mode
+				continue
+			}
 		}
 		if isRegtest := account.CoinCode == coinpkg.CodeRBTC; isRegtest != backend.arguments.Regtest() {
 			// Don't load regtest accounts when running normally, nor mainnet accounts when running
@@ -117,7 +117,7 @@ func (backend *Backend) SupportedCoins(keystore keystore.Keystore) []coinpkg.Cod
 	allCoins := []coinpkg.Code{
 		coinpkg.CodeBTC, coinpkg.CodeTBTC, coinpkg.CodeRBTC,
 		coinpkg.CodeLTC, coinpkg.CodeTLTC,
-		coinpkg.CodeETH, coinpkg.CodeTETH, coinpkg.CodeRETH, coinpkg.CodeGOETH,
+		coinpkg.CodeETH, coinpkg.CodeGOETH,
 	}
 	var availableCoins []coinpkg.Code
 	for _, coinCode := range allCoins {
@@ -222,7 +222,7 @@ func (backend *Backend) createAndPersistAccountConfig(
 			},
 			accountsConfig,
 		)
-	case coinpkg.CodeETH, coinpkg.CodeRETH, coinpkg.CodeTETH, coinpkg.CodeGOETH:
+	case coinpkg.CodeETH, coinpkg.CodeGOETH:
 		bip44Coin := "1'"
 		if coinCode == coinpkg.CodeETH {
 			bip44Coin = "60'"
@@ -674,7 +674,7 @@ func (backend *Backend) persistDefaultAccountConfigs(keystore keystore.Keystore,
 				}
 			}
 		} else {
-			for _, coinCode := range []coinpkg.Code{coinpkg.CodeTBTC, coinpkg.CodeTLTC, coinpkg.CodeTETH, coinpkg.CodeRETH, coinpkg.CodeGOETH} {
+			for _, coinCode := range []coinpkg.Code{coinpkg.CodeTBTC, coinpkg.CodeTLTC, coinpkg.CodeGOETH} {
 				if backend.config.AppConfig().Backend.DeprecatedCoinActive(coinCode) {
 					if _, err := backend.createAndPersistAccountConfig(coinCode, 0, "", keystore, nil, accountsConfig); err != nil {
 						return err