+++ backend/coins/eth/account_test.go
@@ -77,7 +77,7 @@ func newAccount(t *testing.T) *Account {
 			return 0, nil
 		},
 	}
-	coin := NewCoin(client, coin.CodeTETH, "Ropsten", "TETH", "TETH", params.RopstenChainConfig, "", nil, nil)
+	coin := NewCoin(client, coin.CodeGOETH, "Goerli", "GOETH", "GOETH", params.GoerliChainConfig, "", nil, nil)
 	acct := NewAccount(
 		&accounts.AccountConfig{
 			Code:                  "accountcode",