+++ backend/rates/gecko.go
@@ -45,8 +45,6 @@ var (
 		"tbtc":  "bitcoin",
 		"rbtc":  "bitcoin",
 		"tltc":  "litecoin",
-		"teth":  "ethereum",
-		"reth":  "ethereum",
 		"goeth": "ethereum",
 		// ERC20 tokens as used in the backend.
 		// Frontend and app config use unprefixed name, without "eth-erc20-".