+++ frontends/web/src/api/bitbox02.ts
@@ -96,3 +96,16 @@ export const verifyAttestation = (
 ): Promise<boolean | null> => {
   return apiGet(`devices/bitbox02/${deviceID}/attestation`);
 };
+
+export const checkBackup = (
+  deviceID: string,
+  silent: boolean,
+): Promise<string> => {
+  return apiPost(`devices/bitbox02/${deviceID}/backups/check`, { silent })
+    .then((response: FailResponse | (SuccessResponse & { backupID: string; })) => {
+      if (response.success) {
+        return response.backupID;
+      }
+      throw response;
+    });
+};